Move polynomial evaluators to polynomial/
