+++ include/eve/function/diff/horner.hpp
@@ -8,4 +8,4 @@
 #pragma once
 
 #include <eve/function/horner.hpp>
-#include <eve/module/real/polynomials/function/diff/generic/horner.hpp>
+#include <eve/module/real/polynomial/function/diff/generic/horner.hpp>
