+++ include/eve/function/diff/legendre.hpp
@@ -8,4 +8,4 @@
 #pragma once
 
 #include <eve/function/legendre.hpp>
-#include <eve/module/real/math/function/diff/generic/legendre.hpp>
+#include <eve/module/real/polynomial/function/diff/generic/legendre.hpp>
