+++ include/eve/function/hermite.hpp
@@ -14,4 +14,4 @@ namespace eve
   EVE_MAKE_CALLABLE(hermite_, hermite);
 }
 
-#include <eve/module/real/math/function/regular/generic/hermite.hpp>
+#include <eve/module/real/polynomial/function/regular/generic/hermite.hpp>
