+++ include/eve/function/numeric/horner.hpp
@@ -8,4 +8,4 @@
 #pragma once
 
 #include <eve/function/horner.hpp>
-#include <eve/module/real/polynomials/function/numeric/generic/horner.hpp>
+#include <eve/module/real/polynomial/function/numeric/generic/horner.hpp>
