+++ include/eve/function/pedantic/horner.hpp
@@ -8,4 +8,4 @@
 #pragma once
 
 #include <eve/function/horner.hpp>
-#include <eve/module/real/polynomials/function/pedantic/generic/horner.hpp>
+#include <eve/module/real/polynomial/function/pedantic/generic/horner.hpp>
