+++ include/eve/module/real/polynomial/function/diff/generic/gegenbauer.hpp
