+++ include/eve/module/real/polynomial/function/numeric/generic/horner.hpp
@@ -9,23 +9,24 @@
 
 #include <eve/concept/compatible.hpp>
 #include <eve/concept/value.hpp>
-#include <eve/function/regular.hpp>
 #include <eve/constant/one.hpp>
-#include <eve/module/real/polynomials/detail/horner_impl.hpp>
+#include <eve/function/numeric/fma.hpp>
+#include <eve/module/real/polynomial/detail/horner_impl.hpp>
 
 namespace eve::detail
 {
   //================================================================================================
   //== Horner with iterators
   //================================================================================================
-
   template<value T0, std::input_iterator IT>
   EVE_FORCEINLINE constexpr auto horner_(EVE_SUPPORTS(cpu_)
+                                        , numeric_type const &
                                         , T0 xx, IT const & first, IT const & last) noexcept
   requires ((compatible_values<T0, typename std::iterator_traits<IT>::value_type>))
   {
-    return detail::horner_impl(regular_type(), xx, first, last);
+    return detail::horner_impl(numeric_type(), xx, first, last);
   }
+
 
   //================================================================================================
   //== Horner with iterators and leading unitary coefficient
@@ -33,12 +34,13 @@ namespace eve::detail
 
   template<value T0, std::input_iterator IT>
   EVE_FORCEINLINE constexpr auto horner_(EVE_SUPPORTS(cpu_)
+                                        , numeric_type const &
                                         , T0 xx
                                         , callable_one_ const &
                                         , IT const & first, IT const & last) noexcept
   requires ((compatible_values<T0, typename std::iterator_traits<IT>::value_type>))
   {
-    return detail::horner_impl(regular_type(), xx, one, first, last);
+    return detail::horner_impl(numeric_type(), xx, one, first, last);
   }
 
   //================================================================================================
@@ -46,10 +48,11 @@ namespace eve::detail
   //================================================================================================
   template<value T0, range R>
   EVE_FORCEINLINE constexpr auto horner_(EVE_SUPPORTS(cpu_)
+                                        ,  numeric_type const &
                                         , T0 xx, R const & r) noexcept
   requires (compatible_values<T0, typename R::value_type> && (!simd_value<R>))
   {
-    return detail::horner_impl(regular_type(), xx, r);
+    return detail::horner_impl(numeric_type(), xx, r);
   }
 
   //================================================================================================
@@ -57,22 +60,25 @@ namespace eve::detail
   //================================================================================================
   template<value T0, range R>
   EVE_FORCEINLINE constexpr auto horner_(EVE_SUPPORTS(cpu_)
+                                        ,  numeric_type const &
                                         , T0 xx
                                         , callable_one_ const &
                                         , R const & r) noexcept
   requires (compatible_values<T0, typename R::value_type> && (!simd_value<R>))
   {
-    return detail::horner_impl(regular_type(), xx, one, r);
+    return detail::horner_impl(numeric_type(), xx, one, r);
   }
 
   //================================================================================================
-  //== N parameters (((..(a*x+b)*x+c)*x + ..)..)
+  //== Horner variadic N parameters (((..(a*x+b)*x+c)*x + ..)..)
   //================================================================================================
-
-  template<value T0, value ...Ts>
-  EVE_FORCEINLINE constexpr auto horner_(EVE_SUPPORTS(cpu_), T0 x, Ts... args) noexcept
+  template< value T0
+          , value ...Ts>
+  EVE_FORCEINLINE constexpr auto horner_(EVE_SUPPORTS(cpu_)
+                                        , numeric_type const &
+                                        , T0 x, Ts... args) noexcept
   {
-    return horner_impl(regular_type(), x, args...);
+    return horner_impl(numeric_type(), x, args...);
   }
 
   //================================================================================================
@@ -81,9 +87,9 @@ namespace eve::detail
 
   template<value T0, value ...Ts>
   EVE_FORCEINLINE constexpr auto horner_(EVE_SUPPORTS(cpu_)
+                                        , numeric_type const &
                                         , T0 x, callable_one_ const &, Ts... args) noexcept
   {
-    return horner_impl(regular_type(), x, one, args...);
+    return horner_impl(numeric_type(), x, one, args...);
   }
-
 }
