+++ include/eve/module/real/polynomial/function/regular/generic/gegenbauer.hpp
