+++ include/eve/module/real/polynomial/function/regular/generic/hermite.hpp
