+++ include/eve/module/real/polynomial/function/regular/generic/legendre.hpp
